Move Sidebar to `@/components/ui`
