+++ src/components/layout/index.ts
@@ -4,5 +4,4 @@ export * from './Header';
 export * from './Main';
 export * from './OptionsDrawer';
 export * from './OptionsSidebar';
-export * from './Sidebar';
 export * from './ThemeStyle';
