+++ src/components/ui/Sidebar/index.ts
@@ -1,0 +1,1 @@
+export * from './Sidebar';
